Thentos.Adhocracy3.Backend.Api.Simple

Description

This is an implementation of git@github.com:liqdadhocracy3.git:docssourceapi/authentication_api.rst

Documentation

data A3Resource a

Constructors

A3Resource

Fields

mPath :: Maybe Path

mContentType :: Maybe ContentType

mData :: Maybe a

newtype A3UserNoPass

Constructors

A3UserNoPass

Fields

fromA3UserNoPass :: UserFormData

newtype A3UserWithPass

Constructors

A3UserWithPass

Fields

fromA3UserWithPass :: UserFormData

data ActivationRequest

Constructors

ActivationRequest ConfirmationToken

type Api = ThentosApi :<|> (("js" :> Api) :<|> ServiceProxy)

data ContentType

Constructors

CTUser

data LoginRequest

Constructors

LoginByName UserName UserPass

LoginByEmail UserEmail UserPass

data PasswordResetRequest

Constructors

PasswordResetRequest Path UserPass

newtype Path

Constructors

Path

Fields

fromPath :: ST

data RequestResult

Reponse to a login or similar request.

Note: this generic form is useful for parsing a result from A3, but you should never *return* a RequestError because then the status code won't be set correctly. Instead, throw a GenericA3Error or, even better, define and throw a custom error.

Constructors

RequestSuccess Path ThentosSessionToken

RequestError A3ErrorMessage

type ThentosApi = ("principals" :> ("users" :> (ReqBody `[JSON]` A3UserWithPass :> Post200 `[JSON]` TypedPathWithCacheControl))) :<|> (("activate_account" :> (ReqBody `[JSON]` ActivationRequest :> Post200 `[JSON]` RequestResult)) :<|> (("login_username" :> (ReqBody `[JSON]` LoginRequest :> Post200 `[JSON]` RequestResult)) :<|> (("login_email" :> (ReqBody `[JSON]` LoginRequest :> Post200 `[JSON]` RequestResult)) :<|> ("password_reset" :> (ReqBody `[JSON]` PasswordResetRequest :> Post200 `[JSON]` RequestResult)))))

Note: login_username and login_email have identical behavior. In particular, it is not an error to send username and password to /login_email. This makes implementing all sides of the protocol a lot easier without sacrificing security.

data TypedPath

Similar to A3Resource, but tailored for cases where path and content_type are present and data is absent (or irrelevant).

Constructors

TypedPath

Fields

tpPath :: Path

tpContentType :: ContentType

data TypedPathWithCacheControl

The A3 backend replies to POST request not just with a typed path, but also sends a listing of created/modified etc. resources along to allow the A3 frontend to update its cache.

Constructors

TypedPathWithCacheControl

Fields

tpccTypedPath :: TypedPath

tpccChangedDescendants :: [Path]

tpccCreated :: [Path]

tpccModified :: [Path]

tpccRemoved :: [Path]

a3corsPolicy :: CorsPolicy

a3ProxyAdapter :: ProxyAdapter ThentosA3Error

A3-specific ProxyAdapter.

runBackend :: HttpConfig -> ActionState -> IO ()

serveApi :: Manager -> HttpConfig -> ActionState -> Application

thentosApi :: ActionState -> Server ThentosApi